+++ jgclark.EventHelpers/src/eventsToNotes.js
@@ -1,7 +1,7 @@
 // @flow
 // ------------------------------------------------------------------------------------
 // Command to bring calendar events into notes
-// v0.3.0, 4.8.2021
+// v0.3.1, 6.8.2021
 // @jgclark, with additions by @dwertheimer
 // ------------------------------------------------------------------------------------
 
@@ -10,6 +10,7 @@ import {
   toLocaleShortTime,
   stringReplace,
   getTagParams,
+  dateStringFromCalendarFilename,
 } from '../../helperFunctions'
 
 import { getOrMakeConfigurationSection } from '../../nmn.Templates/src/configuration'
@@ -20,7 +21,7 @@ const DEFAULT_EVENTS_OPTIONS = `  events: {
     addEventID: false,  // whether to add an [[event:ID]] internal link when creating an event from a time block
     processedTagName: "#event_created",   // optional tag to add after making a time block an event
     removeTimeBlocksWhenProcessed: true,  // whether to remove time block after making an event from it
-    todaysEventsHeading: "### Events today",  // optional heading to put before list of today's events
+    eventsHeading: "### Events today",  // optional heading to put before list of today's events
     addMatchingEvents: {   // match events with string on left, and then the string on the right is the template for how to insert this event (see README for details)
       "#meeting": "### *|TITLE|* (*|START|*)",
       "#webinar": "### *|TITLE|* (*|START|*)",
@@ -31,7 +32,7 @@ const DEFAULT_EVENTS_OPTIONS = `  events: {
   },
 `
 
-let pref_todaysEventsHeading: string = '### Events today'
+let pref_eventsHeading: string = '### Events today'
 let pref_addMatchingEvents: ?{ [string]: mixed } = null
 let pref_locale: string = 'en-US' // default setting as a backup
 let pref_timeOptions = { hour: '2-digit', minute: '2-digit', hour12: false } // default setting as a backup
@@ -50,20 +51,16 @@ async function getEventsSettings(): Promise<void> {
     await showMessage(`Couldn't find 'events' settings in _configuration note.`)
     return
   }
-  console.log(
-    `\tFound 'events' settings in _configuration note.\nConfig=\n${JSON.stringify(
-      eventsConfig,
-    )}`,
-  )
+  console.log(`\tFound 'events' settings in _configuration note.`)
 
   // now get settings we need
   if (
     eventsConfig.todaysEventsHeading != null &&
     typeof eventsConfig.todaysEventsHeading === 'string'
   ) {
-    pref_todaysEventsHeading = eventsConfig.todaysEventsHeading
+    pref_eventsHeading = eventsConfig.todaysEventsHeading
   }
-  // console.log(pref_todaysEventsHeading)
+  // console.log(pref_eventsHeading)
   if (eventsConfig.addMatchingEvents != null) {
     // $FlowFixMe
     pref_addMatchingEvents = eventsConfig.addMatchingEvents
@@ -82,10 +79,25 @@ async function getEventsSettings(): Promise<void> {
   console.log(`\tEnd of getEventsSettings()`)
 }
 
+async function getEventsForDay(dateStr: string): Promise<Array<TCalendarItem>> {
+  const y = parseInt(dateStr.slice(0,4))
+  const m = parseInt(dateStr.slice(4,6))
+  const d = parseInt(dateStr.slice(6,8))
+  const startOfDay = Calendar.dateFrom(y, m, d, 0, 0, 0)
+  const endOfDay = Calendar.dateFrom(y, m, d, 23, 59, 59)
+  console.log(`  ${startOfDay.toString()} - ${endOfDay.toString()}`)
+  const eA: Array<TCalendarItem> = await Calendar.eventsBetween(startOfDay, endOfDay)
+  console.log(`\tFound ${eA.length} events`)
+  return eA
+}
+
 //------------------------------------------------------------------------------
 // Return MD list of today's events
-export async function listTodaysEvents(paramString?: string): Promise<string> {
-  console.log(`\nlistTodaysEvents: paramString=${String(paramString)}`)
+export async function listDaysEvents(
+  dateStr: string,
+  paramString?: string
+): Promise<string> {
+  console.log(`\nlistDaysEvents for ${dateStr} with paramString=${String(paramString)}`)
 
   // Get config settings from Template folder _configuration note
   await getEventsSettings()
@@ -103,10 +115,12 @@ export async function listTodaysEvents(paramString?: string): Promise<string> {
 
   console.log(`\toutput template: '${template}' and '${allday}'`)
 
-  const eA: Array<TCalendarItem> = await Calendar.eventsToday()
-  console.log(`\tFound ${eA.length} events (including possible dupes)`)
-  const outputArray: Array<string> = []
+  // When it was simply getting *today's* events:
+  // const eA: Array<TCalendarItem> = await Calendar.eventsToday()
+  // Now get the start and end date for the date of the open daily note
+  const eA: Array<TCalendarItem> = await getEventsForDay(dateStr)
 
+  const outputArray: Array<string> = []
   let lastEventStr = '' // keep duplicates from multiple calendars out
   for (const e of eA) {
     // console.log(`      for e: ${e.title}: ${JSON.stringify(e)}`)
@@ -139,8 +153,8 @@ export async function listTodaysEvents(paramString?: string): Promise<string> {
         lastEventStr = thisEventStr
       }
   }
-  if (pref_todaysEventsHeading !== '') {
-    outputArray.unshift(pref_todaysEventsHeading)
+  if (pref_eventsHeading !== '') {
+    outputArray.unshift(pref_eventsHeading)
   }
   const output = outputArray.join('\n')
   console.log(output)
@@ -150,28 +164,32 @@ export async function listTodaysEvents(paramString?: string): Promise<string> {
 //------------------------------------------------------------------------------
 // Insert list of today's events at cursor positions
 // This is called by UI.
-export async function insertListTodaysEvents(params: ?string): Promise<void> {
-  if (Editor.note == null) {
-    await showMessage('Please run again with a note open.')
+export async function insertDaysEvents(
+  paramString: ?string
+): Promise<void> {
+  if (Editor.note == null || Editor.type !== 'Calendar') {
+    await showMessage('Please run again with a calendar note open.')
     return
   }
 
-  // Get list of events happening today
-  let output: string = await listTodaysEvents(params || '')
-  output += output.length === 0 ? '\nnone\n' : '\n'
-  Editor.insertTextAtCursor(output)
+  // Get list of events happening on the day of the open note
+  // $FlowIgnore[incompatible-call]
+  const dateStr = dateStringFromCalendarFilename(Editor.filename)
+  let output: string = await listDaysEvents(dateStr, paramString || '')
+  output += (output.length === 0) ? '\nnone\n' : '\n'
+  Editor.insertTextAtCursor(`::${output}::`)
 }
 
 //------------------------------------------------------------------------------
-// Return string list of matching events in today's note, from list in keys of
-// pref_addMatchingEvents.
-// Prepend any with value of the values in pref_addMatchingEvents.
-export async function listMatchingTodaysEvents(
+// Return string list of matching events in the current day's note, from list 
+// in keys of pref_addMatchingEvents. Apply template before returning.
+export async function listMatchingDaysEvents(
+  dateStr: string,
   /*eslint-disable */
   paramString: ?string, // NB: the parameter isn't currently used, but is provided for future expansion.
   /*eslint-enable */
 ): Promise<string> {
-  console.log(`\nalistMatchingTodaysEvents:`)
+  console.log(`\nlistMatchingDaysEvents for date ${dateStr}:`)
   // Get config settings from Template folder _configuration note
   await getEventsSettings()
   if (pref_addMatchingEvents == null) {
@@ -186,11 +204,13 @@ export async function listMatchingTodaysEvents(
   console.log(
     `\tFrom settings found ${textToMatchA.length} match strings to look for`,
   )
-  const eA: Array<TCalendarItem> = await Calendar.eventsToday()
+
+  const eA: Array<TCalendarItem> = await getEventsForDay(dateStr)
 
   const outputArray: Array<string> = []
   for (const e of eA) {
     for (let i = 0; i < textToMatchA.length; i++) {
+      // FIXME: de-dupe events
       const m = textToMatchA[i]
       const template = templateA[i]
       if (e.title.match(m)) {
@@ -209,24 +229,26 @@ export async function listMatchingTodaysEvents(
       }
     }
   }
-  //     let outputLine = `${textToPrepend[i]}${e.title}`
-  //     if (!e.isAllDay) {
-  //       outputLine += ` (${toLocaleShortTime(e.date)})`
-  //     }
-  //     outputArray.push(outputLine)
-  //   }
-  // }
   const output = outputArray.join('\n')
   console.log(output)
   return output
 }
 
 //------------------------------------------------------------------------------
-// Add matching events to today's note.
-// This is called by UI.
-export async function insertMatchingTodaysEvents(): Promise<void> {
-  console.log(`\ninsertMatchingTodaysEvents:`)
+// Insert list of matching events in the current day's note, from list 
+// in keys of pref_addMatchingEvents. Apply template too.
+export async function insertMatchingDaysEvents(
+  paramString: ?string,
+): Promise<void> {
+  if (Editor.note == null || Editor.type !== 'Calendar') {
+    await showMessage('Please run again with a calendar note open.')
+    return
+  }
+  // $FlowIgnore
+  const dateStr = dateStringFromCalendarFilename(Editor.filename)
+  console.log(`\ninsertMatchingDaysEvents for date ${dateStr}:`)
+
   // Get config settings from Template folder _configuration note
-  const output = await listMatchingTodaysEvents('')
-  Editor.insertTextAtCursor(output)
+  const output = await listMatchingDaysEvents(dateStr, paramString || '')
+  Editor.insertTextAtCursor(`::${output}::`)
 }